requirements.txt
python-telegram-bot==20.7
motor==3.3.2
python-dotenv==1.0.1
